+++ controllers/routers/site.js
@@ -4,30 +4,30 @@ const { ConfigsModel, MediaModel, PostsModel } = require('../../models/');
 
 
 // controllers
-const admin = {};
+const site = {};
 
-admin.main = {
-  GET: function admin_main_GET(req, res, next) {
+site.main = {
+  GET: function site_main_GET(req, res, next) {
     return next();
   },
 };
 
-admin.configs = {
-  GET: function admin_configs_GET(req, res, next) {
+site.configs = {
+  GET: function site_configs_GET(req, res, next) {
     res.locals.$$VIEW.configs = JSON.parse(process.env.$WEBSITE_CONFIGS);
     return next();
   },
-  PATCH: async function admin_configs_PATCH(req, res) {
+  PATCH: async function site_configs_PATCH(req, res) {
     await ConfigsModel.updateConfigs(req.body.configs);                                                                 // tofix: pickup updated variables to avoid injections
     return res.redirect('back');
   },
 };
 
-admin.upload = {                                                                                                        // todo: to be integrated in profile and media manager
-  GET: function admin_upload_GET(req, res, next) {
+site.upload = {                                                                                                        // todo: to be integrated in profile and media manager
+  GET: function site_upload_GET(req, res, next) {
     return next();
   },
-  POST: [_M_.parseMultipart({ fileSize: 25 * 1048576 }), async function admin_upload_POST(req, res) {
+  POST: [_M_.parseMultipart({ fileSize: 25 * 1048576 }), async function site_upload_POST(req, res) {
     if (req.body.busboySlip.mes.length > 0) req.body.busboySlip.mes.forEach(mes => req.flash('error', mes));
     if (req.body.busboySlip.raw.length > 0) {
       req.body.busboySlip.raw.forEach(medium => { medium.author = req.session.user; });
@@ -38,14 +38,14 @@ admin.upload = {                                                                
   }],
 };
 
-admin.stack = {
-  GET: function admin_stack_GET(req, res, next) {
+site.stack = {
+  GET: function site_stack_GET(req, res, next) {
     const collection = req.params.stackType.toLowerCase();
     if (!['posts', 'media'].includes(collection)) throw new _U_.error.HttpException(404);
     _M_.modifyHTMLTitleTag(collection)(req, res);                                                                       // todo: capitalize
     return _M_.aggregateFetch(collection, { num: 10 })(req, res, next);
   },
-  PATCH: async function admin_stack_PATCH(req, res) {
+  PATCH: async function site_stack_PATCH(req, res) {
     const $update = { $set: {} };
     if (req.body.action === 'restored') $update.$set = { 'state.recycled': false };
     else $update.$set = { [`state.${req.body.action}`]: true };
@@ -56,4 +56,4 @@ admin.stack = {
 
 
 // exports
-module.exports = admin;
+module.exports = site;