+++ routers/html/blog.js
@@ -1,31 +1,31 @@
 /* eslint-disable key-spacing */
-const { Device } = require('../controllers/engines/router');
-const { editor, posts } = require('../controllers/routers/posts');
+const { Device } = require('../../controllers/engines/router');
+const { editor, posts: blog } = require('../../controllers/routers/blog');
 
 
 // device
 const postRouter = new Device([
   {
-    route:        ['/edit', '/edit/new'],
+    route:        ['/edit', '/new'],
     controller:   editor.post,
     setting:      { title: 'New post', template: './__root__/editor', authentication: true },
   },
   {
-    route:        /^\/edit\/([a-f\d]{24})(?:\?.*|\/)?$/i,
-    alias:        '/edit/:canonical',
+    route:        /^\/([a-f\d]{24})\/?\/edit\/?$/i,
+    alias:        '/:canonical/edit',
     controller:   editor.edit,
     setting:      { title: 'Edit post', template: './__root__/editor', authorization: true },
   },
   {
-    route:        /^\/(?![a-f\d]{24})(.+)$/i,
+    route:        /^\/(?![a-f\d]{24}\/?$)([^\/]+)\/?$/i,
     alias:        '/:_id',
-    controller:   posts.show,
-    setting:      { template: './theme/posts/posts' },
+    controller:   blog.show,
+    setting:      { template: './theme/blog/posts' },
   },
   {
     route:        '/',
-    controller:   posts.list,
-    setting:      { template: './theme/posts/', handler: Device.handler.VIEW_POSTS_MULTIPLE },
+    controller:   blog.list,
+    setting:      { template: './theme/blog/', handler: Device.handler.VIEW_POSTS_MULTIPLE },
   },
 ]);