+++ routers/index.js
@@ -5,23 +5,23 @@ const errorHandler = require('../controllers/views/error');
 
 
 const routerIndex = {
-  accountRouter: require('./account'),
-  adminRouter: require('./admin'),
-  homeRouter: require('./home'),
-  postsRouter: require('./posts'),
-  pagesRouter: require('./pages'),
-  APIRouter: require('./api'),
-  seedRouter: require('./seed'),
+  APIRouter: require('./api/api'),
+  initRouter: require('./init'),
+  authRouter: require('./auth'),
+  siteRouter: require('./html/site'),
+  homeRouter: require('./html/home'),
+  blogRouter: require('./html/blog'),
+  pageRouter: require('./html/page'),
 };
 
 
 const routerHub = new Device()
-  .use('/seed', routerIndex.seedRouter)
-  .use('/admin', routerIndex.adminRouter)
+  .use('/init', routerIndex.initRouter)
+  .use('/site', routerIndex.siteRouter)
   .use('/home', routerIndex.homeRouter)
-  .use('/posts', routerIndex.postsRouter)
-  .use('/', routerIndex.accountRouter)
-  .use('/', routerIndex.pagesRouter)
+  .use('/blog', routerIndex.blogRouter)
+  .use('/', routerIndex.authRouter)
+  .use('/', routerIndex.pageRouter)
   .use('*', [_M_.responseHTMLRequest, (req, res, next) => next(new HttpException(404))])
   .use(errorHandler)
   .router;